segregation attempt 5

+++ tensorflow/core/framework/tensor_shape.cc
@@ -187,7 +187,7 @@ void TensorShapeBase<Shape>::InitDims(gtl::ArraySlice<int64> dim_sizes) {
                 "bad overflow check");
   bool large_size = false;
   for (auto s : dim_sizes) {
-    if (s > kMaxSmall) {
+    if (static_cast<size_t>(s) > static_cast<size_t>(kMaxSmall)) {
       large_size = true;
       break;
     }
